+++ tests/task/test_cutOutDonutsBase.py
@@ -422,6 +422,7 @@ class TestCutOutDonutsBase(lsst.utils.tests.TestCase):
             "EFFECTIVE",
             "ENTROPY",
             "PEAK_HEIGHT",
+            "FRAC_BAD_PIX",
             "MJD",
             "BORESIGHT_ROT_ANGLE_RAD",
             "BORESIGHT_PAR_ANGLE_RAD",
@@ -447,6 +448,7 @@ class TestCutOutDonutsBase(lsst.utils.tests.TestCase):
             "EFFECTIVE",
             "ENTROPY",
             "PEAK_HEIGHT",
+            "FRAC_BAD_PIX",
         ]:
             self.assertEqual(
                 len(donutStamps), len(donutStamps.metadata.getArray(measure))
@@ -497,6 +499,7 @@ class TestCutOutDonutsBase(lsst.utils.tests.TestCase):
 
         # Add blend to mask
         stamp.wep_im.blendOffsets = [[-50, -60]]
+        stamp.makeMask(self.task.instConfigFile, self.task.opticalModel)
         sn_dict = self.task.calculateSN(stamp)
         for val in sn_dict.values():
             self.assertFalse(np.isnan(val))
@@ -517,3 +520,19 @@ class TestCutOutDonutsBase(lsst.utils.tests.TestCase):
         )
         infoMsg += "of the image; reducing the amount of donut mask dilation to 99"
         self.assertEqual(infoMsg, cm.output[0])
+
+    def testBadPixelMaskDefinitions(self):
+        # Load test data
+        exposure, donutCatalog = self._getExpAndCatalog(DefocalType.Extra)
+
+        # Flag donut pixels as bad
+        self.config.badPixelMaskDefinitions = ["DONUT"]
+        task = CutOutDonutsBaseTask(config=self.config, name="Flag donut pix as bad")
+        donutStamps = task.cutOutStamps(
+            exposure, donutCatalog, DefocalType.Extra, self.cameraName
+        )
+
+        # Check that all the stamps have "bad" pixels
+        # (because we flagged donut pixels as bad)
+        fracBadPix = np.asarray(donutStamps.metadata.getArray("FRAC_BAD_PIX"))
+        self.assertTrue(np.all(fracBadPix > 0))